% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_samples.R
\name{batch_samples}
\alias{batch_samples}
\title{Randomize samples into balanced batches}
\usage{
batch_samples(
  shipment_manifest,
  api_metadata,
  max_n_per_batch,
  strict_size = FALSE,
  max_full_batches = FALSE,
  vars_to_balance = c("codedsiteid", "randomgroupcode", "sex_psca", "older_than_median"),
  outdir = ".",
  verbose = TRUE,
  max_inner_loop_iter = 1e+05,
  max_outer_loop_iter = NULL,
  overwrite = FALSE,
  balance_strictness = NULL,
  tissue_subset = NULL,
  block_randomization = FALSE,
  separate_batch_files = FALSE
)
}
\arguments{
\item{shipment_manifest}{character vector, path(s) or URL to shipment manifests,
e.g. \code{c('Stanford_ADU830-10060_120720.xlsx', 'Stanford_PED830-10062_120720.xlsx')}.
Can be Excel files (.xlsx), tab-delimited text (.tsv), or comma-separated text (.csv).
Must include a "viallabel" column (case insensitive) to map to \code{api_metadata_csv}.
See details for other expected columns.}

\item{api_metadata}{character vector, path(s) or URL to sample metadata from web API,
e.g. \code{'ADU830-10060.csv', 'PED830-10062.csv'}.
Can be Excel files (.xlsx), tab-delimited text (.tsv), or comma-separated text (.csv).
Must include a "viallabel" column (case insensitive) to map to \code{shipment_manifest_excel}.
See details for other expected columns.}

\item{max_n_per_batch}{integer, max number of samples per batch}

\item{strict_size}{bool, force \emph{all} batches to be as close to \code{max_n_per_batch} as possible.
Most applicable for small batches (e.g. < 20). Default: FALSE}

\item{max_full_batches}{bool, whether to force as many batches as possible to have \emph{exactly} \code{max_n_per_batch} samples.
Default: FALSE}

\item{vars_to_balance}{character vector, force batches to include samples from at
least one level of each of these variables. Must be defined in the column names of
\code{api_metadata_csv} or \code{shipment_manifest_excel} (case insensitive).
Default: \code{c('codedsiteid','randomgroupcode','sex_psca','older_than_median')}}

\item{outdir}{character, path to output directory. Current working directory by default.}

\item{verbose}{bool, whether to print progress messages. Default: TRUE.}

\item{max_inner_loop_iter}{integer, max number of failed attempts to fit all samples
in batches before increasing the number of batches. Decreasing this number
may generate batches more quickly, but the number of samples per batch may not be optimal.}

\item{max_outer_loop_iter}{integer, max number of failed attempts to find optimally
balanced batches before relaxing the stringency of the balance checks.
Decreasing this number may generate batches more quickly, but batches may not
be optimally balanced.}

\item{overwrite}{bool, whether to overwrite existing batching results.
Defaults to current working directory.}

\item{balance_strictness}{integer, initial strictness of balance checks,
with 10 being the strictest and 1 being the most lenient. Defaults to 1 if \code{strict_size = TRUE}
and 9 otherwise. Decreasing this number may generate batches more quickly, but batches may not
be optimally balanced.}

\item{tissue_subset}{character, only specified to run batching for a single tissue.
Must be a value in the 'Sample Type' column of one \code{shipment_manifest_excel}
OR a value in the 'SampleTypeCode' column of one \code{api_metadata_csv}."}

\item{block_randomization}{bool, whether to use block randomization.
Samples within a batch are ordered by individual;
samples within an individual are randomized. This adds an 'injection_order' column.
Default: FALSE}

\item{separate_batch_files}{bool, whether to write separate BLINDED output files per batch.
Default: FALSE}
}
\description{
Given different batching considerations for all MoTrPAC Chemical Analysis
Sites (CAS), this function generates batches of a given size that are
balanced in terms of the specified phenotypic variables. Outputs are written
to a local directory.
}
\details{
This function is currently written for compatibility with MoTrPAC clinical sample
shipment manifests and biospecimen metadata files. It can be further generalized if
there is demand to use this function for other projects.

\strong{Inputs:}

Expected column names in the input (case insensitive):
\itemize{
\item Required: \code{viallabel} in both \code{shipment_manifest} and \code{api_metadata}. Sample-level identifier. Used to merge these files.
\item Required: \code{barcode} in either \code{shipment_manifest} or \code{api_metadata}. Another sample-level identifier. Can be NA.
\item Required: \code{pid} in either \code{shipment_manifest} or \code{api_metadata}. Participant-level identifier.
If there is only one sample per participant, this column can be a copy of \code{viallabel}.
\item Required: Each variable listed in \code{vars_to_balance} must be a column name in either \code{shipment_manifest} or \code{api_metadata}.
\item Required: \code{box} in either \code{shipment_manifest} or \code{api_metadata}. Box identifier to indicate which
box/plate the sample was shipped in.
\item Required: \code{position} in either \code{shipment_manifest} or \code{api_metadata}. Provides the position
of the sample in the box/plate in which it was shipped.
\item Required if \code{tissue_subet} is defined: \code{sampletypecode} in \code{api_metadata} or \code{sample type} in \code{shipment_manifest}.
\item Optional: \code{assay} in either \code{shipment_manifest} or \code{api_metadata}. Define this column if subsets of samples are destined for different assays.
\item Required to generate a heatmap of batch balance: \code{'calculatedage','sex_psca','codedsiteid','randomgroupcode'}
}

\strong{Outputs:}

This function writes outputs to a local directory, specified with the \code{outdir} argument.

\emph{Files}

By default, two files are written for each assay & tissue combination:
\itemize{
\item Blinded batch assignments in the format \code{files/[SAMPLE_TYPE]-samples_BLINDED-batch-assignments.csv}
\item Unblinded batching metadata in the format \code{files/[SAMPLE_TYPE]-samples_UNBLINDED-batch-characteristics.csv}
}

Set \code{separate_batch_files} to TRUE to output separate blinded batch assignment files per batch,
e.g. \code{files/[SAMPLE_TYPE]-samples_BLINDED-batch_3-assignments.csv}.

\emph{Plots}

One heatmap is saved for each assay & tissue combination. This plot includes the number of
individuals and samples per batch as well as the balance across each level of each \code{vars_to_balance}.
These plots should be visually examined to confirm that batches are adequately balanced,
i.e. that numbers are reasonably distributed across each ROW. Note that these plots
are currently only generated if the following columns are included in the input files:
'calculatedage','sex_psca','codedsiteid','randomgroupcode'. If these columns are not
included, then batch-by-level tables are printed to the console for each \code{vars_to_balance}.
}
\examples{
# Get paths to shipment and metadata files.
# Here we use examples provided in this R package. 
# These are examples of real MoTrPAC clinical sample batching files, where 
# PIDs, vial labels, and barcodes were changed to random numbers. 
# Note that both adult and pediatric samples are included in these examples,
# which means pediatric and adult samples are randomized together. 
shipment = system.file("extdata", "shipment-manifest.xlsx", package = "SampleBatchR")
metadata = system.file("extdata", "biospecimen-metadata.csv", package = "SampleBatchR")

# For each tissue type present in the input files,
# generate batches with max 96 samples per batch.
# Save outputs to Desktop.
\dontrun{
batch_samples(
 shipment, 
 metadata, 
 max_n_per_batch = 96,
 outdir = "~/Desktop/batches/all" 
)
}

# Generate batches of max 96 samples for muscle samples only (code 6).
# Save outputs to Desktop.
\dontrun{
batch_samples(
 shipment, 
 metadata, 
 max_n_per_batch = 96,
 outdir = "~/Desktop/batches/muscle",
 tissue_subset = "6"
)
}

# Generate small batches of max 15 samples.
# For small batch sizes, `strict_size` should be set to TRUE, 
# which forces as many batches as possible to have exactly N samples. 
\dontrun{
batch_samples(
 shipment, 
 metadata, 
 max_n_per_batch = 15,
 outdir = "~/Desktop/batches/small",
 strict_size = TRUE, 
 tissue_subset = "6"
)
}

# To force as many batches as possible to have exactly N samples
# for larger batch sizes, e.g., >= 30, set `max_full_batches` to TRUE
\dontrun{
batch_samples(
 shipment, 
 metadata, 
 max_n_per_batch = 88,
 outdir = "~/Desktop/batches/max_full",
 max_full_batches = TRUE, 
 tissue_subset = "6"
)
}

# To add an "injection_order" column based on block randomization,
# set `block_randomization` to TRUE. For block randomization, samples within 
# a batch are ordered by individual; samples within an individual are randomized. 
# Set `overwrite` to TRUE to force overwrite of existing results. 
\dontrun{
batch_samples(
 shipment, 
 metadata, 
 max_n_per_batch = 88,
 outdir = "~/Desktop/batches/block_rand",
 block_randomization = TRUE, 
 tissue_subset = "6",
 overwrite = TRUE
)
}

}
