% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_samples.R
\name{check_batch_balance}
\alias{check_batch_balance}
\title{Check batch balance}
\usage{
check_batch_balance(
  curr_batch_pid,
  strictness,
  max_full,
  balance_vars = c("codedsiteid", "randomgroupcode", "sex_psca"),
  verbose = TRUE
)
}
\arguments{
\item{curr_batch_pid}{data table of batches}

\item{strictness}{integer, initial strictness of balance checks,
with 10 being the strictest and 1 being the most lenient}

\item{max_full}{bool, whether to force as many batches as possible to have \emph{exactly} \code{max_n_per_batch} samples}

\item{balance_vars}{character vector of variables for which to check balance}

\item{verbose}{bool, whether to print progress messages}
}
\value{
a named list of two items:
\describe{
\item{\code{success}}{bool, whether batches are balanced}
\item{\code{failed}}{character vector of \code{balance_vars} that were not balanced at the specified strictness}
}
}
\description{
Internal function.
Given a data table of batches, check if the batch is
balanced given the \code{strictness} parameter.
}
\details{
TODO
}
