% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize.R
\name{id_feasible_batch_sizes}
\alias{id_feasible_batch_sizes}
\title{ID feasible batch sizes}
\usage{
id_feasible_batch_sizes(curr_batch_pid, max_n_per_batch, max_full)
}
\arguments{
\item{curr_batch_pid}{data table of input samples}

\item{max_n_per_batch}{integer, max number of samples per batch}

\item{max_full}{bool, whether to force as many batches as possible to have \emph{exactly} \code{max_n_per_batch} samples}
}
\value{
vector of batch sizes
}
\description{
Internal function.
If \link{id_optimal_batch_sizes} fails, identify the minimum possible number of
batches and their ideal sizes. Only used if \code{strict_size} is TRUE.
}
