% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize.R
\name{id_optimal_batch_sizes}
\alias{id_optimal_batch_sizes}
\title{ID optimal batch sizes}
\usage{
id_optimal_batch_sizes(curr_batch_pid, max_n_per_batch, max_full)
}
\arguments{
\item{curr_batch_pid}{data table of input samples}

\item{max_n_per_batch}{integer, max number of samples per batch}

\item{max_full}{bool, whether to force as many batches as possible to have \emph{exactly} \code{max_n_per_batch} samples}
}
\value{
list of optimal batch sizes
}
\description{
Internal function. Given inputs, determine how many samples should be in each batch,
stipulating that all samples from the same patient stay in the same batch.
}
\examples{
TODO
}
