% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_samples.R
\name{plot_batches}
\alias{plot_batches}
\title{Plot batch characteristics}
\usage{
plot_batches(batches, curr_batch, outdir, b)
}
\arguments{
\item{batches}{batching output}

\item{curr_batch}{batching input}

\item{outdir}{character, path to output directory}

\item{b}{character, name of samples being batched}
}
\value{
pheatmap object
}
\description{
Plot a heatmap of the numbers of samples per level per balance variable across batches.
Well-balanced batches are indicated by relatively similar numbers across each ROW of the plot.
This function saves the plot locally and returns the \code{\link[pheatmap:pheatmap]{pheatmap::pheatmap()}} object.
}
\details{
Required columns in \code{batches} include 'calculatedage','sex_psca','codedsiteid','randomgroupcode','batch','N'.
Required columns in \code{curr_batch} include 'codedsiteid','randomgroupcode'.
}
\seealso{
\code{\link[=batch_samples]{batch_samples()}}
}
