% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize.R
\name{randomize}
\alias{randomize}
\title{Randomize samples into balanced batches}
\usage{
randomize(
  shipment_manifest_excel,
  api_metadata_csv,
  max_n_per_batch,
  strict_size = FALSE,
  max_full_batches = FALSE,
  vars_to_balance = c("codedsiteid", "randomgroupcode", "sex_psca", "older_than_40"),
  outdir = ".",
  verbose = TRUE,
  max_inner_loop_iter = 1e+06,
  max_outer_loop_iter = NULL,
  overwrite = FALSE,
  balance_strictness = NULL,
  tissue_subset = NULL,
  block_randomization = FALSE,
  separate_batch_files = FALSE
)
}
\arguments{
\item{shipment_manifest_excel}{character vector, path(s) to shipment manifest Excel files,
e.g. \code{c('Stanford_ADU830-10060_120720.xlsx', 'Stanford_PED830-10062_120720.xlsx')}}

\item{api_metadata_csv}{character vector, path(s) to sample metadata from web API,
e.g. \code{'ADU830-10060.csv', 'PED830-10062.csv'}}

\item{max_n_per_batch}{integer, max number of samples per batch}

\item{strict_size}{bool, force \emph{all} batches to be as close to \code{max_n_per_batch} as possible.
Most applicable for small batches (e.g. < 20).}

\item{max_full_batches}{bool, whether to force as many batches as possible to have \emph{exactly} \code{max_n_per_batch} samples}

\item{vars_to_balance}{character vector, force batches to include samples from at
least one level of each of these variables. Must be defined in \code{api_metadata_csv}.
Defaults to \code{c('codedsiteid','randomgroupcode','sex_psca','older_than_40')}.}

\item{outdir}{character, path to output directory}

\item{verbose}{bool, whether to print progress messages}

\item{max_inner_loop_iter}{integer, max number of failed attempts to fit all samples
in batches before increasing the number of batches}

\item{max_outer_loop_iter}{integer, max number of failed attempts to find optimally
balanced batches before relaxing the stringency of the balance checks}

\item{overwrite}{bool, whether to overwrite existing batching results}

\item{balance_strictness}{integer, initial strictness of balance checks,
with 10 being the strictest and 1 being the most lenient. Defaults to 1 if \code{strict_size = TRUE}
and 9 otherwise.}

\item{tissue_subset}{character, only specified to run batching for a single tissue.
Must be a value in the 'Sample Type' column of one \code{shipment_manifest_excel}
OR a value in the 'SampleTypeCode' column of one \code{api_metadata_csv}."}

\item{block_randomization}{bool, whether to use block randomization.
Samples within a batch are ordered by individual;
samples within an individual are randomized. This adds an 'injection_order' column.}

\item{separate_batch_files}{bool, whether to write separate BLINDED output files per batch}
}
\description{
TODO
}
\details{
TODO
}
\examples{
input = paste0("https://raw.githubusercontent.com/BennyStrobes/Watershed/",
     "master/example_data/watershed_example_data.txt")
     
xlsxFile <- system.file("extdata", "readTest.xlsx", package = "openxlsx")
df1 <- read.xlsx(xlsxFile = xlsxFile, sheet = 1, skipEmptyRows = FALSE)

}
