% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_samples.R
\name{read_input}
\alias{read_input}
\title{Read in input files}
\usage{
read_input(path)
}
\arguments{
\item{path}{character, path or URL to file, e.g. 'Stanford_ADU830-10060_120720.xlsx'}
}
\value{
data table
}
\description{
Internal function used to read in .xlsx, .tsv, or .csv files. "Box"/"box" columns are coerced to character.
}
