% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize.R
\name{read_shipment}
\alias{read_shipment}
\title{Read in shipment manifests}
\usage{
read_shipment(path)
}
\arguments{
\item{path}{character, path to shipment manifest Excel file, e.g. 'Stanford_ADU830-10060_120720.xlsx'}
}
\value{
data table
}
\description{
Internal function used to read in shipment Excel file with specified column types
}
