% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleBatchR-package.R
\docType{package}
\name{SampleBatchR-package}
\alias{SampleBatchR}
\alias{SampleBatchR-package}
\title{SampleBatchR: Create well-balanced batches of experimental samples}
\description{
Make well-balanced batches of MoTrPAC human samples in terms of clinical site, intervention group, age, and sex.
}
\author{
\strong{Maintainer}: Nicole Gay \email{nicole.r.gay@gmail.com}

}
\keyword{internal}
