% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize.R
\name{make_random_batches_not_strict}
\alias{make_random_batches_not_strict}
\title{Make large batches with flexible size}
\usage{
make_random_batches_not_strict(
  curr_batch_pid,
  max_n_per_batch,
  balance_strictness,
  b,
  max_full,
  max_inner_loop_iter,
  max_outer_loop_iter,
  balance_vars = c("codedsiteid", "randomgroupcode", "sex_psca"),
  verbose = T
)
}
\arguments{
\item{curr_batch_pid}{data table of input samples}

\item{max_n_per_batch}{integer, max number of samples per batch}

\item{balance_strictness}{integer, initial strictness of balance checks,
with 10 being the strictest and 1 being the most lenient}

\item{b}{character, sample type description}

\item{max_full}{bool, whether to force as many batches as possible to have \emph{exactly} \code{max_n_per_batch} samples}

\item{max_inner_loop_iter}{integer, max number of failed attempts to fit all samples
in batches before increasing the number of batches}

\item{max_outer_loop_iter}{integer, max number of failed attempts to find optimally
balanced batches before relaxing the stringency of the balance checks}

\item{balance_vars}{character vector, force batches to include samples from at
least one level of each of these variables}

\item{verbose}{bool, whether to print progress messages}
}
\value{
data table of batch assignments or NULL if batching failed
}
\description{
Internal function. TODO.
}
\details{
TODO
}
